+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 
 ## Next (v0.7)
 
+### Breaking
+- Leading and trailing blank lines are ignored by default. instead of kept as blank expressions
+- `ignore_trailing_blanks` renamed to `keep_trailing_blanks`
+- `ignore_blank_lines` renamed to `ignore_all_blanks`
+
 ### New
+- `keep_leading_blanks` compile option
 - `/note` alias for `/text`[^note]
 
 [^note]: I’ve caught myself using `/note` instead of `/text` far too many times, so at this point I think it’s worth adding in the alias :P